Add files via upload
app start activity page that serves for the user to choose whether to continue to use the application or not. On the app start activity page there are two buttons, namely logout to exit the application and the next button to the home menu which is useful for continuing to the next page.